Add files via upload
This program is suppose to simulate of a professor’s office hour that is one hour long. This program also tracks the amount of time a students spends with the professor, the wait time of a student that visits, and the time the professor spends past the designated time. It then outputs the averages of each value over a span of 100 simulations. The program also accounts for students with priority seeing the professor first and keeps track of which student asks which topics during the office hours. In order to keep track of the students and the topics, MAPS and MULTIMAPS stores the data then displays the data both on the terminal and the separate text file called "StudentList&Topics.txt".